requirements.txt
annotated-types==0.7.0
anyio==4.6.0
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.0
h11==0.14.0
httptools==0.6.1
idna==3.10
joblib==1.4.2
numpy==2.0.2
nvidia-nccl-cu12==2.23.4
pandas==2.2.3
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.11
pytz==2024.2
PyYAML==6.0.2
scikit-learn==1.5.2
scipy==1.13.1
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
xgboost==2.1.1
